+++ src/main/java/teammates/common/util/retry/RetryableTaskThrows.java
@@ -1,4 +1,4 @@
-package teammates.test.driver.retry;
+package teammates.common.util.retry;
 
 /**
  * Abstract implementation of a {@link Retryable} task, for easy extending through anonymous classes.
