+++ src/test/java/teammates/test/cases/browsertests/InstructorSubmissionAdjustmentUiTest.java
@@ -13,10 +13,10 @@ import teammates.common.datatransfer.attributes.StudentAttributes;
 import teammates.common.util.AppUrl;
 import teammates.common.util.Const;
 import teammates.common.util.ThreadHelper;
+import teammates.common.util.retry.RetryableTaskReturns;
 import teammates.test.driver.BackDoor;
 import teammates.test.driver.Priority;
 import teammates.test.driver.TestProperties;
-import teammates.test.driver.retry.RetryableTaskReturns;
 import teammates.test.pageobjects.InstructorCourseEnrollPage;
 
 /**
