--- src/test/java/teammates/test/driver/retry/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains infrastructure and helpers needed for retrying arbitrary code in tests.
- */
-package teammates.test.driver.retry;
