+++ src/test/java/teammates/test/pageobjects/AppPage.java
@@ -30,12 +30,12 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import teammates.common.util.Const;
 import teammates.common.util.ThreadHelper;
 import teammates.common.util.Url;
+import teammates.common.util.retry.RetryManager;
+import teammates.common.util.retry.RetryableTaskReturnsThrows;
 import teammates.test.driver.AssertHelper;
 import teammates.test.driver.FileHelper;
 import teammates.test.driver.HtmlHelper;
 import teammates.test.driver.TestProperties;
-import teammates.test.driver.retry.RetryManager;
-import teammates.test.driver.retry.RetryableTaskReturnsThrows;
 
 /**
  * An abstract class that represents a browser-loaded page of the app and
