+++ src/test/java/teammates/test/pageobjects/InstructorCourseJoinConfirmationPage.java
@@ -4,7 +4,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
 import teammates.common.util.Const;
-import teammates.test.driver.retry.RetryableTask;
+import teammates.common.util.retry.RetryableTask;
 
 public class InstructorCourseJoinConfirmationPage extends AppPage {
     @FindBy(id = "button_confirm")
